moved file for firestore user repo and data source
